[NanaZip.Core] Synchronize the 7-Zip ZS implementations to https://github.com/mcmilk/7-Zip-zstd/tree/e4f17d368421568ba9613987d2de8a574a26a472.

+++ NanaZip.Codecs/FastLZMA2/fl2_compress.c
@@ -581,7 +581,10 @@ FL2LIB_API size_t FL2LIB_CALL FL2_compressCCtx(FL2_CCtx* cctx,
     if (cctx->params.doXXH && !cctx->params.omitProp) {
         XXH32_canonical_t canonical;
         DEBUGLOG(5, "Writing hash");
-        if(end - dstBuf < XXHASH_SIZEOF)
+        // **************** 7-Zip ZS Modification Start ****************
+        //if(end - dstBuf < XXHASH_SIZEOF)
+        if (end - dstBuf < (long)XXHASH_SIZEOF)
+        // **************** 7-Zip ZS Modification End ****************
             return FL2_ERROR(dstSize_tooSmall);
         XXH32_canonicalFromHash(&canonical, XXH32(src, srcSize, 0));
         memcpy(dstBuf, &canonical, XXHASH_SIZEOF);
